working towards separating the Parameter modification behavior from the values so that it is easier to serialize/deserialize just the values. replaced inc/dec with adjust which will allow us to pull value o
replaced inc/dec with adjust
made value and object and made it public.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/bravo/JsonBuilder.java
@@ -47,7 +47,7 @@ public class JsonBuilder {
 
     public void append( Param param ){
         buf.append('{');
-        appendProperty("value", param.value); comma();
+        appendProperty("value", (double)param.value); comma();
         appendProperty("isTrue", param.isTrue); comma();
 //        appendProperty("label",param.label);
 //        appendProperty("units", param.units); comma();
